# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_kangzyz_Doub_backend_internal_shared_response.Envelope:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    type: object
...
  internal_transport_http_auth.ErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    type: object
...
  internal_transport_http_auth.LoginResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
      refreshExpiresAt:
        type: string
      sessionID:
        type: string
      twoFactorChallengeToken:
        type: string
      twoFactorRequired:
        type: boolean
      user:
        $ref: '#/definitions/internal_transport_http_auth.UserResponse'
      verificationMethods:
        items:
          type: string
        type: array
    type: object
...
  internal_transport_http_auth.RefreshTokenResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_auth.LoginResponse'
      errorMsg:
        type: string
    type: object
...
  internal_transport_http_auth.UserResponse:
    properties:
      appearancePreferences:
        type: string
      avatarURL:
        type: string
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      emailBootstrapUsedAt:
        type: string
      emailSource:
        type: string
      emailVerifiedAt:
        type: string
      id:
        type: integer
      initialSecurityRequired:
        type: boolean
      initialUsernameRequired:
        type: boolean
      lastLoginAt:
        type: string
      locale:
        type: string
      mustResetPassword:
        type: boolean
      onboardingCompletedAt:
        type: string
      passwordEnabled:
        type: boolean
      passwordOrigin:
        type: string
      passwordSetAt:
        type: string
      phone:
        type: string
      phoneVerifiedAt:
        type: string
      profilePreferences:
        type: string
      publicID:
        type: string
      role:
        type: string
      status:
        type: string
      subscriptionExpiresAt:
        type: string
      subscriptionPlanID:
        type: integer
      subscriptionPlanName:
        type: string
      subscriptionStatus:
        type: string
      subscriptionTier:
        type: string
      timezone:
        type: string
      twoFactorAvailable:
        type: boolean
      twoFactorEnabled:
        type: boolean
      twoFactorRecoveryCount:
        type: integer
      twoFactorRequired:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
      usernameChangedAt:
        type: string
    type: object
...
  internal_transport_http_memory.ErrorDoc:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    type: object
  internal_transport_http_memory.UpsertMemoryResponse:
    properties:
      saved:
        type: boolean
    type: object
  internal_transport_http_memory.UpsertUserMemoryRequest:
    properties:
      memoryKey:
        maxLength: 128
        type: string
      scope:
        enum:
        - profile
        - preference
        - custom
        type: string
      value:
        maxLength: 10000
        type: string
    required:
    - memoryKey
    - scope
    - value
    type: object
  internal_transport_http_memory.UpsertUserMemoryResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_memory.UpsertMemoryResponse'
      errorMsg:
        type: string
    type: object
  internal_transport_http_memory.UserMemoryListResponseDoc:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_transport_http_memory.UserMemoryResponse'
        type: array
      errorMsg:
        type: string
    type: object
  internal_transport_http_memory.UserMemoryResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      memoryKey:
        type: string
      scope:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
      userID:
        type: integer
      value:
        type: string
    type: object
...
info:
  contact: {}
  description: DOUB Chat 后端 API 文档
  title: DOUB Chat API
  version: 0.1.0
paths:
...
  /admin/settings/embedding/reindex:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kangzyz_Doub_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 触发向量重建（重索引所有 stale/failed 文件）
      tags:
      - admin/settings
  /admin/settings/embedding/runtime:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kangzyz_Doub_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询 Embedding 服务运行状态
      tags:
      - admin/settings
  /admin/settings/embedding/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kangzyz_Doub_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 查询向量索引健康状态
      tags:
      - admin/settings
...
  /auth/refresh:
    post:
      description: 使用 HttpOnly refresh cookie 轮换并签发新的 access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_auth.RefreshTokenResponseDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_transport_http_auth.ErrorDoc'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/internal_transport_http_auth.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_auth.ErrorDoc'
      summary: 刷新访问令牌
      tags:
      - auth
...
  /memories/profile:
    get:
      consumes:
      - application/json
      description: 查询当前用户的长期个性化记忆
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_memory.UserMemoryListResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_memory.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 查询用户个性化记忆
      tags:
      - memory
    put:
      consumes:
      - application/json
      description: 新增或更新当前用户的长期个性化记忆
      parameters:
      - description: 用户记忆参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_memory.UpsertUserMemoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_memory.UpsertUserMemoryResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_memory.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_memory.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 更新用户个性化记忆
      tags:
      - memory
  /memories/profile/{memory_key}:
    delete:
      consumes:
      - application/json
      description: 删除当前用户的指定 key 长期记忆
      parameters:
      - description: 记忆 Key
        in: path
        name: memory_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_memory.UpsertUserMemoryResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_memory.ErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_memory.ErrorDoc'
      security:
      - BearerAuth: []
      summary: 删除用户个性化记忆
      tags:
      - memory
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"